Restructured project
- Seperated the functions to other source/header files for a more
  structured project. closes #6
- Changes were made with CMake in mind. However this is the first time I
  use CMake.
- Updated compilation instructions in README.
