+++ .gitignore
@@ -6,6 +6,7 @@
 *.a
 *.lib
 bin/
+build/
 obj/
 main
